+++ packages/pangea-node-sdk/src/services/index.ts
@@ -13,6 +13,7 @@ import {
 import VaultService from "./vault.js";
 import { FileScanService } from "./file_scan.js";
 import ShareService from "./share.js";
+import SanitizeService from "./sanitize.js";
 
 export default {
   AuditService,
@@ -28,4 +29,5 @@ export default {
   VaultService,
   FileScanService,
   ShareService: ShareService,
+  SanitizeService,
 };